Moved small classes to another file
