+++ src/generator.cpp
@@ -16,243 +16,6 @@ limitations under the License.
 
 #include "libcellml/generator.h"
 
-class Representable
-{
-public:
-    virtual std::string repr() = 0;
-};
-
-class BinaryOperator : public Representable
-{
-public:
-    BinaryOperator() :
-        arg1(std::shared_ptr<Representable>()),
-        arg2(std::shared_ptr<Representable>())
-    {}
-
-    BinaryOperator(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2) :
-        arg1(arg1),
-        arg2(arg2)
-    {}
-
-    std::shared_ptr<Representable> getArg1() const {return arg1;}
-    void setArg1(const std::shared_ptr<Representable> a1) {arg1 = a1;}
-    std::shared_ptr<Representable> getArg2() const {return arg2;}
-    void setArg2(const std::shared_ptr<Representable> a2) {arg2 = a2;}
-
-protected:
-    std::shared_ptr<Representable> arg1;
-    std::shared_ptr<Representable> arg2;
-};
-
-class ArithmeticOperator : public BinaryOperator
-{
-public:
-    ArithmeticOperator(std::string opr) :
-        BinaryOperator(),
-        operatorRepresentation(opr)
-    {}
-
-    ArithmeticOperator(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2,
-            std::string opr) :
-        BinaryOperator(arg1,arg2),
-        operatorRepresentation(opr)
-    {}
-
-    virtual std::string repr() override
-    {
-        std::ostringstream oss;
-        oss << "(" << arg1->repr();
-        oss << " " << operatorRepresentation << " ";
-        oss << arg2->repr() << ")";
-        return oss.str();
-    }
-
-private:
-    std::string operatorRepresentation;
-};
-
-class Addition : public ArithmeticOperator
-{
-public:
-    Addition() :
-        ArithmeticOperator(std::string("+"))
-    {}
-
-    Addition(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2) :
-        ArithmeticOperator(arg1,arg2,std::string("+"))
-    {}
-};
-
-class Subtraction : public ArithmeticOperator
-{
-public:
-    Subtraction() :
-        ArithmeticOperator(std::string("-"))
-    {}
-
-    Subtraction(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2) :
-        ArithmeticOperator(arg1,arg2,std::string("-"))
-    {}
-};
-
-class Multiplication : public ArithmeticOperator
-{
-public:
-    Multiplication() :
-        ArithmeticOperator(std::string("*"))
-    {}
-
-    Multiplication(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2) :
-        ArithmeticOperator(arg1,arg2,std::string("*"))
-    {}
-};
-
-class Division : public ArithmeticOperator
-{
-public:
-    Division() :
-        ArithmeticOperator(std::string("/"))
-    {}
-
-    Division(std::shared_ptr<Representable> arg1,
-            std::shared_ptr<Representable> arg2) :
-        ArithmeticOperator(arg1,arg2,std::string("/"))
-    {}
-};
-
-class Power : public BinaryOperator
-{
-public:
-    Power(std::shared_ptr<Representable> b,
-            std::shared_ptr<Representable> e) :
-        BinaryOperator(b,e)
-    {}
-
-    virtual std::string repr() override
-    {
-        std::ostringstream oss;
-        oss << "std::pow(" << arg1->repr() << ", " << arg2->repr() << ")";
-        return oss.str();
-    }
-};
-
-class UnaryOperator : public Representable
-{
-public:
-    UnaryOperator() :
-        arg(std::shared_ptr<Representable>())
-    {}
-
-    UnaryOperator(std::shared_ptr<Representable> arg) :
-        arg(arg)
-    {}
-
-    std::shared_ptr<Representable> getArg() const {return arg;}
-    void setArg(const std::shared_ptr<Representable> a) {arg = a;}
-
-protected:
-    std::shared_ptr<Representable> arg;
-};
-
-class STDOperator : public UnaryOperator
-{
-public:
-    STDOperator(std::string fun) :
-        UnaryOperator(),
-        fun(fun)
-    {}
-
-    STDOperator(std::shared_ptr<Representable> arg,
-            std::string fun) :
-        UnaryOperator(arg),
-        fun(fun)
-    {}
-
-    virtual std::string repr() override
-    {
-        std::ostringstream oss;
-        oss << "std::" << fun << "(" << arg->repr() << ")";
-        return oss.str();
-    }
-
-private:
-    std::string fun;
-};
-
-class AbsoluteValue : public STDOperator
-{
-public:
-    AbsoluteValue() :
-        STDOperator(std::string("abs"))
-    {}
-
-    AbsoluteValue(std::shared_ptr<Representable> arg) :
-        STDOperator(arg,std::string("abs"))
-    {}
-};
-
-class Sine : public STDOperator
-{
-public:
-    Sine() :
-        STDOperator(std::string("sin"))
-    {}
-
-    Sine(std::shared_ptr<Representable> arg) :
-        STDOperator(arg,std::string("sin"))
-    {}
-};
-
-class Cosine : public STDOperator
-{
-public:
-    Cosine() :
-        STDOperator(std::string("cos"))
-    {}
-
-    Cosine(std::shared_ptr<Representable> arg) :
-        STDOperator(arg,std::string("cos"))
-    {}
-};
-
-class Variable : public Representable
-{
-public:
-    Variable() = default;
-    Variable(std::string name) : name(name) {}
-
-    virtual std::string repr() override
-    {
-        return name;
-    }
-
-private:
-    std::string name;
-};
-
-class Constant : public Representable
-{
-public:
-    Constant() = default;
-    Constant (double val) : value(val) {}
-
-    virtual std::string repr() override
-    {
-        std::ostringstream oss;
-        oss << std::setprecision(16) << value;
-        return oss.str();
-    }
-
-private:
-    double value;
-};
-
 namespace libcellml{
 
 std::string Generator::generateStateAliases()
